refactor: 💡 one concern per agent/lib module
autumnDirectTools shrinks to the resolver alone: token cache lives with
autumnAuth, the metadata cache and leaf MCP address in
autumnToolMetadata, schema slimming in toolSchemaSlim, and
connection_search dedup in discoveredConnectionTools.
